chore: split code into separate files
