+++ src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,1 @@
-{"migrated":{"identifier":"migrated","description":"permissions that were migrated from v1","local":true,"windows":["main"],"permissions":["app:default","event:default","image:default","path:default","menu:default","resources:default","shell:allow-open","tray:default","window:default"]}}
+{"migrated":{"identifier":"migrated","description":"permissions that were migrated from v1","local":true,"windows":["main"],"permissions":["app:default","event:default","image:default","path:default","menu:default","resources:default","shell:allow-open","tray:default","window:default","window:allow-start-dragging","window:allow-start-resize-dragging"]}}
